% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVfromVCF.R
\name{MutationsTable}
\alias{MutationsTable}
\title{Compute a mutation table containing useful informations}
\usage{
MutationsTable(vcf, genome, context_length = 3)
}
\arguments{
\item{vcf}{(Mandatory) ExpandedVCF object from which mutations are taken}

\item{genome}{(Mandatory) Genome used to retrieve the sequences}

\item{context_length}{(Mandatory, default: 3) The length of output sequences}
}
\value{
mut_table data.frame containing the results
}
\description{
This function take as input a ExpandedVCF object, a genome and a context
length then it provides as output a dataframe containing for each mutation
chromosome, strand, position, type of SNV, the reference sequence and the
alternative sequence. Note that the mutation from G to A is the same as
C to T on the reverse strand, which implies a redundancy that is
addressed by converting the mutation types obtained for mutations of REF
bases A and G to their respective reverse complements.
Thus all mutation types are reported such that they have C or T as the
mutated REF base. The mutations which are kept as they are are on the
positive strand and the mutations for which the reverse complement is
computed are on the negative strand.
}
\examples{

data(vcf_chr1_example)
genome <- PreprocessGenome("hg38", vcf = vcf_chr1_example)
mut_table <- MutationsTable(vcf_chr1_example, genome, context_length = 3)

}
