% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVfromVCF.R
\name{PreprocessVCF}
\alias{PreprocessVCF}
\title{Preprocess a VCF file}
\usage{
PreprocessVCF(vcf)
}
\arguments{
\item{vcf}{(Mandatory) Path to input file.}
}
\value{
filtered_vcf An ExpandedVCF object.
}
\description{
This function loads a VCF file and filters it keeping only Single Nucleotide
Variants, it extends the CollapsedVCF object in order to have in the ALT
field only one variant per line then it return the ExpandedVCF object for
further analysis.
}
\examples{

path <- system.file("example-data", "example.vcf", package="SNVfromVCF")
vcf <- PreprocessVCF(path)

}
