% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVfromVCF-package.R
\docType{package}
\name{SNVfromVCF-package}
\alias{SNVfromVCF}
\alias{SNVfromVCF-package}
\title{SNVfromVCF: A simple tool for SNP mutation type determination}
\description{
A simple tool for SNP mutation type determination. From a VCF file and a reference genome it extracts the reference sequence of length specified by the user and its corresponding variant then there is a function that counts the occurrencies of each SNV and provides a simple plot visualization. This package is designed for analizing human variants.
}
\author{
\strong{Maintainer}: Kevin Del Gaudio \email{kevindelgaudiokdg@gmail.com}

}
\keyword{internal}
