% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVfromVCF.R
\name{TypeCounts}
\alias{TypeCounts}
\title{Count the occurrence of each SNV and plot it.}
\usage{
TypeCounts(mut_table, plot = FALSE)
}
\arguments{
\item{mut_table}{(Mandatory) A dataframe generated by the MutationsTable
function}

\item{plot}{(Mandatory, default: FALSE) A bool that computes the plot if true}
}
\value{
count_table A table object with the counts of the mutation types.
}
\description{
This function counts the occurrences of each SNV and, if requested by the
user, it shows a barplot summarizing the counts.
}
\examples{

data(vcf_chr1_example)
genome <- PreprocessGenome("hg38", vcf = vcf_chr1_example)
mut_table <- MutationsTable(vcf_chr1_example, genome, context_length = 3)
count_table <- TypeCounts(mut_table, plot = TRUE)
#compute count table and show the barplot

}
