% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_chr1_example.R
\docType{data}
\name{vcf_chr1_example}
\alias{vcf_chr1_example}
\title{A VCF containing chromosome 1 variants}
\format{
An object of class ExpandedVCF
}
\usage{
data(vcf_chr1_example)
}
\value{
An ExpandedVCF containing mutation data of chromosome 1
}
\description{
Variantcalling data from an experiment performed on chromosome 1 of a patient
}
\examples{
data(vcf_chr1_example)
ref <- ref(vcf_chr1_example)
}
\keyword{datasets}
